pyproject.toml
[project]
name = "viby"
version = "0.2.2"
description = "Viby vibes everything"
authors = [
    { name = "JohanLi233", email = "li_zhonghan@qq.com" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.12",
    "fastmcp>=2.3.4",
    "flatlatex>=0.15",
    "numpy>=2.2.5",
    "openai>=1.76.2",
    "pocketflow>=0.0.2",
    "prompt-toolkit>=3.0.51",
    "pyperclip>=1.9.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "rich>=14.0.0",
    "sentence-transformers>=4.1.0",
    "typer>=0.15.4",
]

[project.optional-dependencies]
dev = [
    "black>=23.1.0",
    "isort>=5.12.0",
    "mypy>=1.0.1",
]
test = [
    "pytest>=7.3.1",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.10.0",
    "coverage>=7.2.2",
]

[tool.setuptools.packages.find]
include = ["viby*"]
exclude = ["assets*"]

[project.scripts]
yb = "viby.cli.main:entry_point"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
markers = [
    "unit: 单元测试",
    "integration: 集成测试",
    "e2e: 端到端测试",
    "performance: 性能测试",
]

[tool.coverage.run]
source = ["viby"]
omit = ["tests/*", "**/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
    "raise ImportError",
]
